+++ Kernel/Modules/AdminProcessManagementActivityDialog.pm
@@ -80,7 +80,7 @@ sub Run {
     }
 
     my $DynamicFieldList = $Kernel::OM->Get('Kernel::System::DynamicField')->DynamicFieldList(
-        ObjectType => [ 'Ticket', 'Article' ],
+        ObjectType => ['Ticket'],
         ResultType => 'HASH',
     );
 
@@ -672,7 +672,7 @@ sub _ShowEdit {
     }
 
     my $DynamicFieldList = $Kernel::OM->Get('Kernel::System::DynamicField')->DynamicFieldList(
-        ObjectType => [ 'Ticket', 'Article' ],
+        ObjectType => ['Ticket'],
         ResultType => 'HASH',
     );
 
@@ -888,6 +888,33 @@ sub _ShowEdit {
         Translation   => 1,
         Class         => 'Modernize',
     );
+
+    my %TimeUnitsSelectionList = (
+        0 => 'Do not show Field',
+        2 => 'Show Field As Mandatory',
+    );
+
+    if ( !$ConfigObject->Get('Ticket::Frontend::NeedAccountedTime') ) {
+        $TimeUnitsSelectionList{1} = 'Show Field';
+    }
+
+    # create TimeUnits selection
+    if ( $ConfigObject->Get('Ticket::Frontend::AccountTime') ) {
+
+        $Param{TimeUnitsSelection} = $LayoutObject->BuildSelection(
+            Data          => \%TimeUnitsSelectionList,
+            SelectedValue => 0,
+            Name          => 'TimeUnits',
+            ID            => 'TimeUnits',
+            Translation   => 1,
+            Class         => 'Modernize',
+        );
+
+        $LayoutObject->Block(
+            Name => 'TimeUnitsContainer',
+            Data => \%Param,
+        );
+    }
 
     # extract parameters from config
     $Param{DescriptionShort} = $Param{ActivityDialogData}->{Config}->{DescriptionShort};